requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
appdirs==1.4.3            # via black
attrs==19.1.0             # via black
black==19.3b0
click==7.0                # via black
django-tastypie==0.14.2
django==2.2.4
psycopg2-binary==2.8.3
python-dateutil==2.8.0    # via django-tastypie
python-dotenv==0.10.3
python-mimeparse==1.6.0   # via django-tastypie
pytz==2019.2              # via django
six==1.12.0               # via python-dateutil
sqlparse==0.3.0           # via django
toml==0.10.0              # via black
